change add profile style
