+++ controllers/cartController.js
@@ -53,7 +53,7 @@ const cartController = {
         name: 'cart'
       }
     })
-      .spread(function (cart, created) {
+      .spread(function(cart, created) {
         return CartItem.findOrCreate({
           where: {
             CartId: cart.id,
@@ -68,7 +68,7 @@ const cartController = {
             sugar: req.body.sugar
           }
         })
-          .spread(function (cartItem, created) {
+          .spread(function(cartItem, created) {
             return cartItem
               .update({
                 quantity: cartItem.quantity
@@ -84,6 +84,7 @@ const cartController = {
                     req.session.cartId = cart.id
                     req.session.cart_number = totalQuantity
                     return req.session.save(() => {
+                      req.flash('success_messages', '已加入購物車')
                       return res.redirect('back')
                     })
                   })
@@ -145,7 +146,7 @@ const cartController = {
       default: {
         name: 'cart'
       }
-    }).spread(function (cart, created) {
+    }).spread(function(cart, created) {
       console.log(req.body.amount)
       return CartItem.findOrCreate({
         where: {
@@ -160,7 +161,7 @@ const cartController = {
           ice: req.body.ice,
           sugar: req.body.sugar
         }
-      }).spread(function (cartItem, created) {
+      }).spread(function(cartItem, created) {
         return cartItem
           .update({
             quantity: cartItem.quantity
