+++ package.json
@@ -20,6 +20,7 @@
     "express": "^4.17.1",
     "express-handlebars": "^3.1.0",
     "express-session": "^1.17.0",
+    "express-validator": "^6.2.0",
     "faker": "^4.1.0",
     "imgur-node-api": "^0.1.0",
     "method-override": "^3.0.0",
