+++ public/stylesheets/product.css
@@ -68,7 +68,7 @@ h6 {
   height: 100%;
 }
 
-button {
+.button-style {
   background-color: #013b4f !important;
   color: #bc955c !important;
 }
@@ -104,4 +104,27 @@ span {
 
 .fb_iframe_widget {
   height: 100%;
+}
+
+.ratings {
+  position: relative;
+  vertical-align: middle;
+  display: inline-block;
+  color: #ddd;
+  /*背景星星顏色*/
+  overflow: hidden;
+  font-size: 30px;
+  /*調整字體大小可放大縮小星星*/
+  text-shadow: 0px 1px 0 #999;
+}
+
+.full_star {
+  /*調整寬度可變更星等*/
+  position: absolute;
+  left: 0;
+  top: 0;
+  white-space: nowrap;
+  overflow: hidden;
+  color: #d56a16;
+  /*前景星星顏色*/
 }
