+++ public/stylesheets/profile.css
@@ -1,0 +1,72 @@
+.col-md-12 {
+  padding: 0;
+}
+
+.card-header {
+  background-color: #013b4f;
+  color: #bc955c;
+}
+
+a,
+a:hover {
+  color: #bc955c;
+  text-decoration: none;
+}
+
+.a-button {
+  width: 48%;
+}
+
+.btn-primary,
+.btn-primary:hover {
+  background-color: #013b4f;
+  color: #bc955c;
+  border-color: #013b4f;
+}
+
+.edit {
+  border-color: #bc955c;
+}
+
+.edit:hover {
+  background-color: #bc955c;
+  color: #013b4f;
+}
+
+.link {
+  width: 100%;
+  height: 100%;
+}
+
+.btn-danger {
+  background-color: #013b4f;
+  border-color: #bc955c;
+}
+
+.btn-danger:hover,
+.btn-danger:focus {
+  background-color: #bc955c;
+  color: #013b4f;
+  border-color: #bc955c;
+}
+
+.btn-danger:not(:disabled):not(.disabled).active,
+.btn-danger:not(:disabled):not(.disabled):active,
+.show > .btn-danger.dropdown-toggle,
+.btn-primary:not(:disabled):not(.disabled).active,
+.btn-primary:not(:disabled):not(.disabled):active,
+.show > .btn-primary.dropdown-toggle {
+  background-color: #bc955c;
+  color: #013b4f;
+  border-color: #bc955c;
+}
+
+.btn-primary.focus,
+.btn-primary:focus {
+  box-shadow: 0 0 0.2rem #bd9b69;
+}
+
+.btn.focus,
+.btn:focus {
+  outline: #bc955c;
+}
